feat(examples): add guarded data-container GPU example
Add the published top-level runnable example for the data-container
and GPU-transfer walkthrough, with a guide-local forwarder to keep the
topic-directory path in sync.

Also expand smoke coverage for the documented CPU-only and Warp-backed
round-trip behavior so the example stays validated end to end.

Closes #1223

ADW-ID: e9abde5e
