+++ particula/gpu/tests/data_containers_example_test.py
@@ -20,13 +20,14 @@ EXAMPLE_PATH = (
     / "Examples"
     / "data_containers_and_gpu_foundations.py"
 )
-IMPLEMENTATION_PATH = (
+GUIDE_PATH = (
     Path(__file__).resolve().parents[3]
     / "docs"
     / "Examples"
     / "Data_Containers"
     / "data_containers_and_gpu_foundations.py"
 )
+EXAMPLES_ROOT = EXAMPLE_PATH.parent
 
 
 def _load_module(module_name: str, module_path: Path) -> types.ModuleType:
@@ -43,20 +44,24 @@ def _load_module(module_name: str, module_path: Path) -> types.ModuleType:
 
 
 @pytest.fixture
-def example_impl_module() -> types.ModuleType:
-    """Load the topic-directory example implementation module."""
+def example_module(monkeypatch: pytest.MonkeyPatch) -> types.ModuleType:
+    """Load the published top-level example module."""
+    monkeypatch.syspath_prepend(str(EXAMPLES_ROOT))
+    sys.modules.pop("data_containers_and_gpu_foundations", None)
     return _load_module(
-        "data_containers_and_gpu_foundations_impl_test",
-        IMPLEMENTATION_PATH,
+        "data_containers_and_gpu_foundations_test",
+        EXAMPLE_PATH,
     )
 
 
 @pytest.fixture
-def example_wrapper_module() -> types.ModuleType:
-    """Load the published top-level wrapper module."""
+def guide_module(monkeypatch: pytest.MonkeyPatch) -> types.ModuleType:
+    """Load the guide-local forwarding module."""
+    monkeypatch.syspath_prepend(str(EXAMPLES_ROOT))
+    sys.modules.pop("data_containers_and_gpu_foundations", None)
     return _load_module(
-        "data_containers_and_gpu_foundations_wrapper_test",
-        EXAMPLE_PATH,
+        "data_containers_and_gpu_foundations_guide_test",
+        GUIDE_PATH,
     )
 
 
@@ -75,10 +80,10 @@ def _run_example(*, force_no_warp: bool) -> subprocess.CompletedProcess[str]:
 
 
 def test_build_particle_data_returns_documented_shapes(
-    example_impl_module: types.ModuleType,
+    example_module: types.ModuleType,
 ) -> None:
     """Test particle example data matches the documented single-box schema."""
-    particle_data = example_impl_module._build_particle_data()
+    particle_data = example_module._build_particle_data()
 
     assert particle_data.masses.shape == (1, 2, 2)
     assert particle_data.concentration.shape == (1, 2)
@@ -89,10 +94,10 @@ def test_build_particle_data_returns_documented_shapes(
 
 
 def test_build_gas_data_returns_documented_shapes_and_names(
-    example_impl_module: types.ModuleType,
+    example_module: types.ModuleType,
 ) -> None:
     """Test gas example data matches the documented single-box schema."""
-    gas_data = example_impl_module._build_gas_data()
+    gas_data = example_module._build_gas_data()
 
     assert gas_data.name == ["Water", "H2SO4"]
     assert gas_data.molar_mass.shape == (2,)
@@ -103,24 +108,24 @@ def test_build_gas_data_returns_documented_shapes_and_names(
 
 def test_warp_enabled_honors_force_no_warp_environment(
     monkeypatch: pytest.MonkeyPatch,
-    example_impl_module: types.ModuleType,
+    example_module: types.ModuleType,
 ) -> None:
     """Test the force-no-warp environment variable disables Warp transfers."""
-    monkeypatch.setattr(example_impl_module, "WARP_AVAILABLE", True)
+    monkeypatch.setattr(example_module, "WARP_AVAILABLE", True)
     monkeypatch.setenv("PARTICULA_EXAMPLE_FORCE_NO_WARP", "1")
 
-    assert example_impl_module._warp_enabled() is False
+    assert example_module._warp_enabled() is False
 
 
 def test_run_example_reports_cpu_only_message_when_warp_disabled(
     monkeypatch: pytest.MonkeyPatch,
-    example_impl_module: types.ModuleType,
+    example_module: types.ModuleType,
 ) -> None:
     """Test run_example returns the documented CPU-only success message."""
-    monkeypatch.setattr(example_impl_module, "WARP_AVAILABLE", False)
+    monkeypatch.setattr(example_module, "WARP_AVAILABLE", False)
     monkeypatch.delenv("PARTICULA_EXAMPLE_FORCE_NO_WARP", raising=False)
 
-    output = example_impl_module.run_example()
+    output = example_module.run_example()
 
     assert len(output) == 3
     assert output[0].startswith("ParticleData constructed:")
@@ -128,42 +133,41 @@ def test_run_example_reports_cpu_only_message_when_warp_disabled(
     assert "completed without Warp" in output[2]
 
 
-def test_wrapper_main_prints_example_output(
+def test_example_main_prints_example_output(
     monkeypatch: pytest.MonkeyPatch,
     capsys: pytest.CaptureFixture[str],
-    example_wrapper_module: types.ModuleType,
+    example_module: types.ModuleType,
 ) -> None:
-    """Test the published wrapper prints the forwarded example output."""
+    """Test the published example prints the documented output."""
     monkeypatch.setenv("PARTICULA_EXAMPLE_FORCE_NO_WARP", "1")
 
-    example_wrapper_module.main()
+    example_module.main()
 
     captured = capsys.readouterr()
     assert "ParticleData constructed:" in captured.out
     assert "Warp-backed transfers are optional" in captured.out
 
 
-def test_implementation_main_prints_example_output(
+def test_guide_main_prints_example_output(
     monkeypatch: pytest.MonkeyPatch,
     capsys: pytest.CaptureFixture[str],
-    example_impl_module: types.ModuleType,
+    guide_module: types.ModuleType,
 ) -> None:
-    """Test the implementation main prints the example lines."""
-    monkeypatch.setattr(example_impl_module, "WARP_AVAILABLE", False)
-    monkeypatch.delenv("PARTICULA_EXAMPLE_FORCE_NO_WARP", raising=False)
+    """Test the guide-local module delegates to the canonical example."""
+    monkeypatch.setenv("PARTICULA_EXAMPLE_FORCE_NO_WARP", "1")
 
-    example_impl_module.main()
+    guide_module.main()
 
     captured = capsys.readouterr()
     assert "ParticleData constructed:" in captured.out
     assert "completed without Warp" in captured.out
 
 
-def test_wrapper_runs_as_main_entrypoint(
+def test_example_runs_as_main_entrypoint(
     monkeypatch: pytest.MonkeyPatch,
     capsys: pytest.CaptureFixture[str],
 ) -> None:
-    """Test the published wrapper executes successfully as __main__."""
+    """Test the published example executes successfully as __main__."""
     monkeypatch.setenv("PARTICULA_EXAMPLE_FORCE_NO_WARP", "1")
 
     runpy.run_path(str(EXAMPLE_PATH), run_name="__main__")
@@ -173,14 +177,14 @@ def test_wrapper_runs_as_main_entrypoint(
     assert "Warp-backed transfers are optional" in captured.out
 
 
-def test_implementation_runs_as_main_entrypoint(
+def test_guide_runs_as_main_entrypoint(
     monkeypatch: pytest.MonkeyPatch,
     capsys: pytest.CaptureFixture[str],
 ) -> None:
-    """Test the implementation executes successfully as __main__."""
+    """Test the guide-local forwarding module executes successfully."""
     monkeypatch.setenv("PARTICULA_EXAMPLE_FORCE_NO_WARP", "1")
 
-    runpy.run_path(str(IMPLEMENTATION_PATH), run_name="__main__")
+    runpy.run_path(str(GUIDE_PATH), run_name="__main__")
 
     captured = capsys.readouterr()
     assert "ParticleData constructed:" in captured.out
@@ -194,6 +198,19 @@ def test_example_non_warp_path_reports_cpu_success() -> None:
     assert "ParticleData constructed:" in result.stdout
     assert "GasData constructed:" in result.stdout
     assert "Warp-backed transfers are optional" in result.stdout
+
+
+def test_guide_module_re_exports_canonical_run_example(
+    monkeypatch: pytest.MonkeyPatch,
+    guide_module: types.ModuleType,
+) -> None:
+    """Test the guide module re-exports the canonical run_example helper."""
+    monkeypatch.setenv("PARTICULA_EXAMPLE_FORCE_NO_WARP", "1")
+
+    output = guide_module.run_example()
+
+    assert output[0].startswith("ParticleData constructed:")
+    assert "completed without Warp" in output[-1]
 
 
 @pytest.mark.skipif(not WARP_AVAILABLE, reason="Warp is not available")
@@ -211,13 +228,13 @@ def test_example_warp_path_reports_round_trip_shapes_and_names() -> None:
 @pytest.mark.skipif(not WARP_AVAILABLE, reason="Warp is not available")
 def test_run_example_warp_path_reports_round_trip_shapes_and_names(
     monkeypatch: pytest.MonkeyPatch,
-    example_impl_module: types.ModuleType,
+    example_module: types.ModuleType,
 ) -> None:
     """Test run_example reports the documented Warp round-trip details."""
-    monkeypatch.setattr(example_impl_module, "WARP_AVAILABLE", True)
+    monkeypatch.setattr(example_module, "WARP_AVAILABLE", True)
     monkeypatch.delenv("PARTICULA_EXAMPLE_FORCE_NO_WARP", raising=False)
 
-    output = example_impl_module.run_example()
+    output = example_module.run_example()
 
     assert any("Warp particle round trip:" in line for line in output)
     assert any("restored_masses=(1, 2, 2)" in line for line in output)
